refactor(files): extract FileListService from FileController (thin controller, fat service)
